audit: kcbVarFreeze pin on int32_t for LP64 safety
aetFreeze events store a single bool-as-int in the GG variable part,
sized via sizeof(long) — 4 bytes on x86 and Win64 LLP64, 8 on LP64.
That would shift the on-disk size of every aetFreeze entry on Linux/Mac
x64 builds.

Pin the macro to int32_t and convert all 7 read/write locals
(actor.cpp x4, actredit.cpp x1, actor.cpp::AssertValid x1, movie_chomp.cpp x1).

x86 build green; on-disk byte stream unchanged.

+++ inc/actor.h
@@ -291,7 +291,7 @@ const ByteOrderMask kbomAevrot = kbomBmat34;
 #define kcbVarSize (size(BRS))
 #define kcbVarPull (size(Stretch))
 #define kcbVarSnd (size(Sound))
-#define kcbVarFreeze (size(long))
+#define kcbVarFreeze (size(int32_t))
 #define kcbVarMove (size(RoutePoint))
 #define kcbVarTweak (size(RoutePoint))
 #define kcbVarStep (size(BRS))
